copilot-kit/a2a-travel-demo-app/agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a>=0.1.0",
    "a2a-sdk[http-server]>=0.3.20",
    "ag-ui-adk>=0.0.1",
    "fastapi>=0.115.0",
    "google-adk>=0.1.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "litellm>=1.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.30.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
